treewide: make mode unsigned short
The filemode we are tracking fits in an unsigned short.  Many places
that print the mode or otherwise format it into a string expect that
as well.  Make the mode an unsigned short where necessary, so it can
be printed without warnings when "-Wformat-signedness" is turned on.

+++ wt-status.h
@@ -46,7 +46,7 @@ struct wt_status_change_data {
 	int worktree_status;
 	int index_status;
 	int stagemask;
-	int mode_head, mode_index, mode_worktree;
+	unsigned short mode_head, mode_index, mode_worktree;
 	struct object_id oid_head, oid_index;
 	int rename_status;
 	int rename_score;
